refactored time_frequency to convert and notation
