+++ librosa/core/__init__.py
@@ -2,13 +2,14 @@
 # -*- coding: utf-8 -*-
 """ Core IO and DSP functions"""
 
-from .time_frequency import *  # pylint: disable=wildcard-import
+from .convert import *  # pylint: disable=wildcard-import
 from .audio import *  # pylint: disable=wildcard-import
 from .spectrum import *  # pylint: disable=wildcard-import
 from .pitch import *  # pylint: disable=wildcard-import
 from .constantq import *  # pylint: disable=wildcard-import
 from .harmonic import *  # pylint: disable=wildcard-import
 from .fft import *  # pylint: disable=wildcard-import
+from .notation import *  # pylint: disable=wildcard-import
 
 
 __all__ = [_ for _ in dir() if not _.startswith('_')]
